Rename component index files
