+++ libs/design-system/src/index.ts
@@ -1,12 +1,9 @@
 export { Attribution } from './lib/Components/Attribution/Attribution';
-export { Heading, HeadingProps } from './lib/Components/Heading/Heading';
+export { Heading, HeadingProps } from './lib/Components/Heading';
 export {
   InternalSidebar,
   TEMP_SIDEBAR_CONFIG,
-} from './lib/Components/InternalSidebar/InternalSidebar';
-export { getHeadingsFromMdxComponent } from './lib/Components/InternalToc/InternalToc';
-export { Link } from './lib/Components/Link/Link';
-export {
-  InputProps,
-  StaticCheckbox,
-} from './lib/Components/StaticCheckbox/StaticCheckbox';
+} from './lib/Components/InternalSidebar';
+export { getHeadingsFromMdxComponent } from './lib/Components/InternalToc';
+export { Link } from './lib/Components/Link';
+export { InputProps, StaticCheckbox } from './lib/Components/StaticCheckbox';
