+++ libs/design-system/src/lib/Components/Heading/Heading.stories.tsx
@@ -1,5 +1,5 @@
 import { Meta, Story } from '@storybook/react';
-import { Heading, HeadingProps } from './Heading';
+import { Heading, HeadingProps } from './';
 
 export default {
   component: Heading,
